Refact and extract components from EditorSidebar
