+++ src/icons/TrashIcon.tsx
@@ -1,0 +1,75 @@
+/* @jsx h */
+import { h } from "preact";
+
+export default function TrashIcon(
+  { class: className, width = 16, height = 16 }: h.JSX.SVGAttributes<
+    SVGElement
+  >,
+) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={width}
+      height={height}
+      class={className}
+      fill="#000000"
+      viewBox="0 0 256 256"
+    >
+      <rect width="256" height="256" fill="none"></rect>
+      <line
+        x1="216"
+        y1="56"
+        x2="40"
+        y2="56"
+        fill="none"
+        stroke="#000000"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="16"
+      >
+      </line>
+      <line
+        x1="104"
+        y1="104"
+        x2="104"
+        y2="168"
+        fill="none"
+        stroke="#000000"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="16"
+      >
+      </line>
+      <line
+        x1="152"
+        y1="104"
+        x2="152"
+        y2="168"
+        fill="none"
+        stroke="#000000"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="16"
+      >
+      </line>
+      <path
+        d="M200,56V208a8,8,0,0,1-8,8H64a8,8,0,0,1-8-8V56"
+        fill="none"
+        stroke="#000000"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="16"
+      >
+      </path>
+      <path
+        d="M168,56V40a16,16,0,0,0-16-16H104A16,16,0,0,0,88,40V56"
+        fill="none"
+        stroke="#000000"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="16"
+      >
+      </path>
+    </svg>
+  );
+}
